cleanup file structure and remove code from main
