+++ backend/src/api/mod.rs
@@ -1,8 +1,10 @@
 mod aur;
 pub mod backend;
 mod build;
+pub mod cusom_file_server;
 #[cfg(feature = "static")]
 pub mod embed;
+pub mod init;
 mod package;
 mod stats;
 mod types;
