+++ backend/src/repo/mod.rs
@@ -1,1 +1,2 @@
+pub mod init;
 pub mod repo;
